Initial Commit
This is a fork of the oscam software [ http://www.streamboard.tv/oscam ]
with a Android.mk for easy compiling with the rest of the platform.

It Compiles. But Does it acutally work? That's the chance you take
because as of this commit I haven't tested this with any hardware

Good Luck.